Use standard gem from Gemfile
I was recently running StandardRB on our codebase locally and noticed
that the output was different to that of CI. Turns out it was happening
because the action just does a gem install standard whilst development
uses the version pinned in the Gemfile.lock. I bumped our version of
standard in development and output was then consistent between
environments.

That was a temporary fix, because as new standard gems are released we
would get the same issue again, but we wouldn't be alerted of that fact
and would get weird failures.

Instead of doing a `gem install` we install the version from our
`Gemfile` which will always be the same version in development.