Install the same Ruby version as in our Gemfile
Because we install the gem from the `Gemfile` the Ruby version specified
there must be what we use to install `standard`. This will cause us
maintenance when we upgrade Ruby, but the good thing is that this check
will fail, and we will be alerted to the fact Ruby version in the
`Dockerfile` needs bumping.